+++ src/flextype/Support/Serializers/Json.php
@@ -39,7 +39,7 @@ class Json
      *
      * @return mixed A JSON string representing the original PHP value
      */
-    public function encode($input, int $options = 0, int $depth = 512): string
+    public function encode($input, int $options = 0, int $depth = 512)
     {
         $options = ($options & self::ESCAPE_UNICODE ? 0 : JSON_UNESCAPED_UNICODE)
             | JSON_UNESCAPED_SLASHES
@@ -66,7 +66,7 @@ class Json
      *
      * @return mixed The JSON converted to a PHP value
      *
-     * @throws ParseException If the JSON is not valid
+     * @throws RuntimeException If the JSON is not valid
      */
     public function decode(string $input, bool $cache = true, bool $assoc = true, int $depth = 512, int $flags = 0)
     {