+++ src/flextype/Support/Serializers/Yaml.php
@@ -80,9 +80,9 @@ class Yaml
      *
      * @return mixed The YAML converted to a PHP value
      *
-     * @throws ParseException If the YAML is not valid
+     * @throws RuntimeException If the YAML is not valid
      */
-    public function decode(string $input, bool $cache = true, int $flags = 0): array
+    public function decode(string $input, bool $cache = true, int $flags = 0)
     {
         $decode = function (string $input, int $flags = 0) {
             // Try native PECL YAML PHP extension first if available.